backend-nexus/requirements.txt
fastapi==0.110.0
uvicorn[standard]==0.27.1
SQLAlchemy>=2.0.36
psycopg[binary]>=3.1.0
psycopg2-binary>=2.9.0
alembic==1.13.1
python-dotenv==1.0.1
pydantic[email]>=2.12.0
passlib[bcrypt]==1.7.4
bcrypt==4.0.1
pytest>=7.0.0
python-jose
httpx
